Fixed crash with empty tag value + refactored AdvancedPoiEditing + changed entity id to long long

+++ Sources/OsmEdit/OAOpenStreetMapLocalUtil.mm
@@ -67,7 +67,7 @@ static const int NON_AMENITY_ID_RIGHT_SHIFT = 7;
     return newEntity;
 }
 
-- (OAEntityInfo *)getEntityInfo:(long)identifier
+- (OAEntityInfo *)getEntityInfo:(long long)identifier
 {
     return nil;
 }
