+++ Sources/OsmEdit/OsmData/OAEntity.h
@@ -40,16 +40,16 @@ typedef NS_ENUM(NSInteger, EOAEntityType)
 
 @interface OAEntity : NSObject <OAEntityProtocol>
 
--(id)initWithId:(long)identifier;
--(id)initWithId:(long)identifier latitude:(double)lat longitude:(double)lon;
--(id)initWithEntity:(OAEntity *)copy identifier:(long)identifier;
+-(id)initWithId:(long long)identifier;
+-(id)initWithId:(long long)identifier latitude:(double)lat longitude:(double)lon;
+-(id)initWithEntity:(OAEntity *)copy identifier:(long long)identifier;
 
 
 -(NSSet<NSString *> *) getChangedTags;
 -(void) setChangedTags:(NSSet<NSString *> *)changedTags;
 -(NSInteger) getModify;
 -(void)setModify:(NSInteger)modify;
--(long) getId;
+-(long long) getId;
 -(double)getLatitude;
 -(double) getLongitude;
 
@@ -85,11 +85,11 @@ typedef NS_ENUM(NSInteger, EOAEntityType)
 
 @interface OAEntityId : NSObject
 
--(id) initWithEntityType:(EOAEntityType)type identifier:(long)identifier;
+-(id) initWithEntityType:(EOAEntityType)type identifier:(long long)identifier;
 +(OAEntityId *) valueOf:(OAEntity *)entity;
 -(NSString *) toNSString;
 -(EOAEntityType) getType;
--(long) getId;
+-(long long) getId;
 -(NSString *) getOsmUrl;
 
 @end
